+++ plugins/postcss-double-position-gradients/src/index.js
@@ -1,4 +1,6 @@
+import postcssProgressiveCustomProperties from '@csstools/postcss-progressive-custom-properties';
 import valueParser from 'postcss-value-parser';
+import { hasSupportsAtRuleAncestor } from './has-supports-at-rule-ancestor';
 
 // whether the value has a lab() or lch() matcher
 const gradientRegExp = /(repeating-)?(conic|linear|radial)-gradient\([\W\w]*\)/i;
@@ -26,13 +28,15 @@ function isNumericNode(node) {
  * @param {{preserve?: boolean}} opts
  * @returns {import('postcss').Plugin}
  */
-function creator(opts) {
-	const preserve = 'preserve' in Object(opts) ? Boolean(opts.preserve) : true;
-
+const basePlugin = (opts) => {
 	return {
 		postcssPlugin: 'postcss-double-position-gradients',
 		Declaration(decl, { result }) {
 			if (!gradientRegExp.test(decl.value)) {
+				return;
+			}
+
+			if (hasSupportsAtRuleAncestor(decl)) {
 				return;
 			}
 
@@ -89,7 +93,7 @@ function creator(opts) {
 			const modifiedValue = valueAST.toString();
 
 			if (modifiedValue !== decl.value) {
-				if (preserve) {
+				if (opts.preserve) {
 					decl.cloneBefore({ value: modifiedValue });
 					return;
 				}
@@ -98,8 +102,34 @@ function creator(opts) {
 			}
 		},
 	};
-}
+};
 
-creator.postcss = true;
+basePlugin.postcss = true;
 
-export default creator;
+/**
+ * Transform double-position gradients in CSS.
+ * @param {{preserve?: boolean}} opts
+ * @returns {import('postcss').Plugin}
+ */
+const postcssPlugin = (opts) => {
+	const options = Object.assign({
+		enableProgressiveCustomProperties: true,
+		preserve: true,
+	}, opts);
+
+	if (options.enableProgressiveCustomProperties && options.preserve) {
+		return {
+			postcssPlugin: 'postcss-double-position-gradients',
+			plugins: [
+				postcssProgressiveCustomProperties(),
+				basePlugin(options),
+			],
+		};
+	}
+
+	return basePlugin(options);
+};
+
+postcssPlugin.postcss = true;
+
+export default postcssPlugin;